requirements.txt
django>=1.8
flake8
python-oauth2==0.7.0
uwsgi
python-social-auth==0.2.7
oauthlib==1.0.3
django-wysiwyg-redactor==0.4.9
pillow==2.9.0
django-formtools==1.0
djangorestframework==3.1.3
mysqlclient==1.3.6
django-oauth-toolkit==0.9.0
pinax-notifications==3.0.0
pymorphy2==0.8
django-filter
django-super-inlines==0.1.4
